+++ CHANGES.md
@@ -10,7 +10,7 @@ Beta Releases
    * Replaced the `positionToEyeWC` field of the `czm_materialInput` struct with its equivalent in eye coordinates, `positionToEyeEC`.
 * Added EllipsoidPrimitive for visualizing ellipsoids and spheres.
 * Added Matrix4.multiplyByTranslation, Matrix4.fromScale, and Matrix3.fromScale.
-* Added new Materials: `RimLighting`, TODO.  See the [Fabric](https://github.com/AnalyticalGraphicsInc/cesium/wiki/Fabric) wiki page.
+* Added new Materials: `RimLighting`, ErosionMaterial`, TODO.  See the [Fabric](https://github.com/AnalyticalGraphicsInc/cesium/wiki/Fabric) wiki page.
 
 ### b8 - xx/xx/2012